requirements.txt
tensorflow==1.14.0
numpy>=1.18.0
nltk==3.4.5
tflearn==0.3.2
bottle>=0.12.18
request>=2.22.0